requirements.txt
streamlit==1.26.0
tensorflow==2.14.0
pillow
numpy